Pretty: a bit of refactoring
