+++ disco.cabal
@@ -278,6 +278,8 @@ library
                        Disco.Module
                        Disco.Parser
                        Disco.Pretty
+                       Disco.Pretty.Monadic
+                       Disco.Pretty.Prec
                        Disco.Property
                        Disco.Desugar
                        Disco.Compile
